refactor(lib): rename `gemini` to `google`
